+++ internal/provider/warehouse_data_source.go
@@ -88,7 +88,7 @@ func (d *warehouseDataSource) Read(ctx context.Context, req datasource.ReadReque
 	}
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Unable to Read Warehouse",
+			fmt.Sprintf("Unable to Read Warehouse (ID: %s)", state.ID.ValueString()),
 			getError(err, body),
 		)